Refactor `monitorVetoSum`
